feat(database): initialize database with users and graphs tables
- Added initial structure for `users` table.
- Modified the `create` function in `routesv1.rs` to handle POST requests and insert data into the specified service's table based on JSON input. This includes type checking, query formatting, and executing the prepared SQL query.
- Added a new `get_all` function in `routesv1.rs` to fetch all rows from the specified service's table and print them.
- Updated account struct in `account.rs` to implement sqlx::FromRow to enable fetching records into the struct directly.

+++ server/src/routesv1/account.rs
@@ -6,9 +6,9 @@ use actix_web::{
 use serde_json::to_value;
 
 /// Account related structs and handlers
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq)]
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize, PartialEq, sqlx::FromRow)]
 pub struct Account {
-    pub id: u32,
+    pub id: i64,
     pub username: String,
     pub email: String,
     pub password: String,
